Basic Character Controller setup with the neccessary variables. Character Movement started with keypresses being passed through update and not a polling event. Created more MIDI files to test. Updated schedule. correct

+++ BeatBreaker/MIDIParse.h
@@ -28,7 +28,7 @@ public:
 
 	// Shift bits
 	uint32_t read_uint32(std::ifstream& t_file);
-	uint32_t read_uint16(std::ifstream& t_file);
+	uint16_t read_uint16(std::ifstream& t_file);
 
 private:
